added logic to add / remove from the db and show when registering / un-registering to a publication  added registeredUser db fields for phone, name, and uuid  added sms and call intent to registered users of the admin's publication

+++ app/src/main/java/com/roa/foodonetv3/activities/GroupsActivity.java
@@ -211,7 +211,7 @@ public class GroupsActivity extends AppCompatActivity implements NavigationView.
     @Override
     public void onNewGroupClick(String groupName){
         /** after a user creates a new group from the dialog, run the service to create the group */
-        Group newGroup = new Group(groupName, CommonMethods.getMyUserID(this),-1);
+        Group newGroup = new Group(groupName, CommonMethods.getMyUserID(this),(long)-1);
         Intent intent = new Intent(this, FoodonetService.class);
         intent.putExtra(ReceiverConstants.ACTION_TYPE, ReceiverConstants.ACTION_ADD_GROUP);
         intent.putExtra(ReceiverConstants.JSON_TO_SEND,newGroup.getAddGroupJson().toString());
@@ -239,7 +239,7 @@ public class GroupsActivity extends AppCompatActivity implements NavigationView.
                             break;
                         case ADMIN_GROUP_TAG:
                             /** pressed on create a new user in a group the user is the admin of */
-                            // TODO: 14/12/2016 currently not working... getting a 404 code
+                            // TODO: 14/12/2016 currently not working... getting a 440 json response
                             Toast.makeText(this, "add new member", Toast.LENGTH_SHORT).show();
                             Intent fabClickIntent = new Intent(ReceiverConstants.BROADCAST_FOODONET);
                             fabClickIntent.putExtra(ReceiverConstants.ACTION_TYPE,ReceiverConstants.ACTION_FAB_CLICK);
